Extract shared device transfer foundation
